requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.85.1
h11==0.14.0
idna==3.4
packaging==21.3
passlib==1.7.4
Pillow==9.3.0
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
pytesseract==0.3.10
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
uvicorn==0.19.0
